requirements.txt
google-cloud-bigquery
confluent-kafka
faker
pandas
psycopg2